refactor: eliminate barrel index.ts files and update consumer imports (#199)
- Deleted unused barrel files: hooks/index.ts, header/index.ts, dialogs/index.ts
- Flattened listing-wizard/types/index.ts → types.ts and constants/index.ts → constants.ts
- Stripped hook re-exports from listing-wizard/index.ts
- Deleted component barrels: auction/index.ts, bidding/index.ts, admin/index.ts, listing-wizard/index.ts
- Split convex/admin/index.ts into convex/admin/queries.ts + convex/admin/mutations.ts
- Updated all consumer import paths to direct file paths
- Updated test mocks to use new direct file paths
