+++ convex/admin/queries.ts
@@ -1,35 +1,41 @@
 /**
- * Admin panel operations - core functionality.
+ * Admin panel query functions.
  *
- * This module contains the main admin operations including bid management,
- * support ticket handling, audit logging, and announcements.
+ * This module consolidates all read-only admin operations:
+ * - Bid monitoring: getRecentBids
+ * - Support: getTickets
+ * - Audit: getAuditLogs
+ * - Announcements: listAnnouncements
  *
- * Specialized operations are organized in sub-modules:
- * - kyc.ts: KYC verification and review
- * - statistics.ts: Dashboard metrics and reporting
+ * Re-exports from specialized sub-modules:
+ * - kyc.ts: getPendingKYC
+ * - statistics.ts: getFinancialStats, getAdminStats, getAnnouncementStats, getSupportStats
+ * - settings.ts: getSystemConfig
  */
 
 import { v } from "convex/values";
 import { paginationOptsValidator } from "convex/server";
 
-import { mutation, query } from "../_generated/server";
-import { requireAdmin, getAuthUser, UnauthorizedError } from "../lib/auth";
-import { logAudit, updateCounter, countQuery } from "../admin_utils";
+import { query } from "../_generated/server";
+import { requireAdmin, UnauthorizedError } from "../lib/auth";
+import { countQuery } from "../admin_utils";
 import { batchFetchReadCounts } from "../notifications";
 import type { Doc, Id } from "../_generated/dataModel";
 
-// --- Re-export specialized modules for backward compatibility ---
-export { getPendingKYC, reviewKYC } from "./kyc";
+// --- Re-export specialized query modules ---
+
+export { getPendingKYC } from "./kyc";
+
 export {
   getFinancialStats,
   getAdminStats,
   getAnnouncementStats,
   getSupportStats,
-  initializeCounters,
 } from "./statistics";
-export { getSystemConfig, updateSystemConfig } from "./settings";
 
-// --- Bid Moderation ---
+export { getSystemConfig } from "./settings";
+
+// --- Bid Monitoring ---
 
 /**
  * Query recent bids across all auctions.
@@ -90,12 +96,10 @@ export const getRecentBids = query({
         };
       }
 
-      // Collect unique auction IDs
       const uniqueAuctionIds = [
         ...new Set(bidsResult.page.map((b) => b.auctionId)),
       ];
 
-      // Fetch all unique auctions in parallel
       const auctionMap = new Map<
         Id<"auctions">,
         Doc<"auctions"> | null | { _error: true }
@@ -159,67 +163,6 @@ export const getRecentBids = query({
       console.error("Critical error in getRecentBids:", err);
       throw err;
     }
-  },
-});
-
-/**
- * Mark a bid as voided and recalculate auction pricing.
- *
- * - Marks the bid as voided
- * - Recalculates the auction's current price to the next highest valid bid
- * - Reverts to starting price if no valid bids exist
- * - Logs the action for audit
- *
- * Only accessible to admin users.
- */
-export const voidBid = mutation({
-  args: { bidId: v.id("bids"), reason: v.string() },
-  returns: v.object({ success: v.boolean() }),
-  handler: async (ctx, args) => {
-    await requireAdmin(ctx);
-
-    const bid = await ctx.db.get(args.bidId);
-    if (!bid) throw new Error("Bid not found");
-    if (bid.status === "voided") return { success: true }; // Already voided
-
-    // Mark as void
-    await ctx.db.patch(args.bidId, { status: "voided" });
-
-    // Recalculate Auction Price (Next highest valid bid)
-    const auction = await ctx.db.get(bid.auctionId);
-    if (!auction) throw new Error("Auction not found");
-
-    const latestValidBid = await ctx.db
-      .query("bids")
-      .withIndex("by_auction", (q) => q.eq("auctionId", bid.auctionId))
-      .filter((q) => q.neq(q.field("status"), "voided"))
-      .filter((q) => q.neq(q.field("_id"), bid._id))
-      .order("desc")
-      .first();
-
-    const newPrice = latestValidBid
-      ? latestValidBid.amount
-      : auction.startingPrice;
-    const newWinnerId = latestValidBid ? latestValidBid.bidderId : null;
-
-    const patchData: { currentPrice: number; winnerId?: string | null } = {
-      currentPrice: newPrice,
-    };
-    if (auction.winnerId !== newWinnerId) {
-      patchData.winnerId = newWinnerId;
-    }
-
-    await ctx.db.patch(bid.auctionId, patchData);
-
-    // Log Action
-    await logAudit(ctx, {
-      action: "VOID_BID",
-      targetId: args.bidId,
-      targetType: "bid",
-      details: `Reason: ${args.reason}. New Price: ${String(newPrice)}${auction.winnerId !== newWinnerId ? `. Winner recalculated to ${String(newWinnerId)}` : ""}`,
-    });
-
-    return { success: true };
   },
 });
 
@@ -307,53 +250,7 @@ export const getTickets = query({
   },
 });
 
-/**
- * Resolve a support ticket with a resolution comment.
- *
- * Marks the ticket as resolved and records the admin who resolved it.
- * Only accessible to admin users.
- */
-export const resolveTicket = mutation({
-  args: { ticketId: v.id("supportTickets"), resolution: v.string() },
-  returns: v.object({ success: v.boolean() }),
-  handler: async (ctx, args) => {
-    await requireAdmin(ctx);
-
-    const authUser = await getAuthUser(ctx);
-    if (!authUser) {
-      throw new Error("Admin identity not found or invalid");
-    }
-
-    const ticket = await ctx.db.get(args.ticketId);
-    if (!ticket) {
-      throw new Error("Ticket not found");
-    }
-
-    if (ticket.status === "resolved") {
-      return { success: true };
-    }
-
-    await ctx.db.patch(args.ticketId, {
-      status: "resolved",
-      updatedAt: Date.now(),
-      resolvedBy: authUser.userId ?? authUser._id,
-    });
-
-    if (ticket.status === "open") {
-      await updateCounter(ctx, "support", "open", -1);
-      await updateCounter(ctx, "support", "resolved", 1);
-    }
-
-    await logAudit(ctx, {
-      action: "RESOLVE_TICKET",
-      targetId: args.ticketId,
-      targetType: "supportTicket",
-      details: JSON.stringify({ resolution: args.resolution }),
-    });
-
-    return { success: true };
-  },
-});
+// --- Audit Logs ---
 
 /**
  * Query audit logs of admin actions.
@@ -412,51 +309,7 @@ export const getAuditLogs = query({
   },
 });
 
-// --- Communication / Announcements ---
-
-/**
- * Create a platform-wide announcement.
- *
- * Broadcasts an announcement to all users as a notification.
- * Only accessible to admin users.
- */
-export const createAnnouncement = mutation({
-  args: { title: v.string(), message: v.string() },
-  returns: v.object({ success: v.boolean() }),
-  handler: async (ctx, args) => {
-    await requireAdmin(ctx);
-
-    const title = args.title.trim();
-    const message = args.message.trim();
-
-    if (title.length === 0 || title.length > 200) {
-      throw new Error("Title must be between 1 and 200 characters");
-    }
-    if (message.length === 0 || message.length > 2000) {
-      throw new Error("Message must be between 1 and 2000 characters");
-    }
-
-    await ctx.db.insert("notifications", {
-      recipientId: "all",
-      type: "info",
-      title,
-      message,
-      isRead: false,
-      createdAt: Date.now(),
-    });
-
-    await updateCounter(ctx, "announcements", "total", 1);
-
-    await logAudit(ctx, {
-      action: "CREATE_ANNOUNCEMENT",
-      targetId: "all",
-      targetType: "announcement",
-      details: title,
-    });
-
-    return { success: true };
-  },
-});
+// --- Announcements ---
 
 /**
  * List all announcements with read counts.
@@ -519,7 +372,6 @@ export const listAnnouncements = query({
       };
     }
 
-    // Parallel fetch read counts using indexed queries
     const announcementIds = announcementsResult.page.map((a) => a._id);
     const readCounts = await batchFetchReadCounts(ctx, announcementIds);
 
@@ -532,79 +384,6 @@ export const listAnnouncements = query({
       ...announcementsResult,
       page,
       totalCount,
-    };
-  },
-});
-
-/**
- * Maintenance mutation to synchronize winnerId with the highest bidder for all auctions.
- *
- * Processes auctions in batches to avoid runtime/memory limits.
- * Returns a cursor if more auctions need processing.
- *
- * Only accessible to admin users.
- */
-export const syncAuctionWinners = mutation({
-  args: {
-    cursor: v.optional(v.string()),
-    batchSize: v.optional(v.number()),
-  },
-  returns: v.object({
-    processed: v.number(),
-    updated: v.number(),
-    continueCursor: v.union(v.string(), v.null()),
-    isDone: v.boolean(),
-  }),
-  handler: async (ctx, args) => {
-    await requireAdmin(ctx);
-
-    const batchSize = Math.max(
-      1,
-      Math.min(Math.floor(args.batchSize ?? 50) || 50, 100)
-    );
-
-    const auctionsQuery = ctx.db.query("auctions");
-
-    // We use the default ordering which is by _creationTime
-    const results = await auctionsQuery.paginate({
-      numItems: batchSize,
-      cursor: args.cursor ?? null,
-    });
-
-    let updatedCount = 0;
-
-    for (const auction of results.page) {
-      const highestBid = await ctx.db
-        .query("bids")
-        .withIndex("by_auction", (q) => q.eq("auctionId", auction._id))
-        .filter((q) => q.neq(q.field("status"), "voided"))
-        .order("desc")
-        .first();
-
-      const currentWinnerId = highestBid ? highestBid.bidderId : null;
-
-      if (auction.winnerId !== currentWinnerId) {
-        await ctx.db.patch(auction._id, {
-          winnerId: currentWinnerId,
-        });
-        updatedCount++;
-      }
-    }
-
-    if (updatedCount > 0) {
-      await logAudit(ctx, {
-        action: "SYNC_AUCTION_WINNERS_BATCH",
-        targetId: "batch",
-        targetType: "auction",
-        details: `Processed batch of ${String(results.page.length)} auctions, updated ${String(updatedCount)} winners.`,
-      });
-    }
-
-    return {
-      processed: results.page.length,
-      updated: updatedCount,
-      continueCursor: results.continueCursor,
-      isDone: results.isDone,
     };
   },
 });
