--- src/components/admin/index.ts
@@ -1,16 +1,0 @@
-export * from "./AdminLayout";
-export * from "./BidMonitor";
-export * from "./StatCard";
-export * from "./SummaryCard";
-export * from "./AdminConnectionError";
-export * from "./ModerationCard";
-export * from "./SettingsCard";
-export * from "./DetailItem";
-export * from "./AuditTab";
-export * from "./FinanceTab";
-export * from "./SupportTab";
-export * from "./EmptyState";
-export * from "./ConditionItem";
-export * from "./EquipmentMetadataEditor";
-export * from "./CategoryManager";
-export * from "./MetadataCatalog";
