--- src/components/auction/index.ts
@@ -1,3 +1,0 @@
-export * from "./AuctionCard";
-export * from "./AuctionCardThumbnail";
-export * from "./AuctionCardPrice";
