--- src/components/bidding/index.ts
@@ -1,3 +1,0 @@
-export * from "./BiddingPanel";
-export * from "./BidForm";
-export * from "./BidHistory";
