--- src/components/header/index.ts
@@ -1,4 +1,0 @@
-export * from "./Header";
-export * from "./MobileMenu";
-export * from "./SearchBar";
-export * from "./UserDropdown";
