--- src/components/listing-wizard/index.ts
@@ -1,12 +1,0 @@
-export * from "./ListingWizard";
-export * from "./StepIndicator";
-export * from "./WizardNavigation";
-export * from "./context/ListingWizardContext";
-export * from "@/hooks/listing-wizard/useListingForm";
-export * from "@/hooks/listing-wizard/useListingMedia";
-export * from "./steps/GeneralInfoStep";
-export * from "./steps/TechnicalSpecsStep";
-export * from "./steps/ConditionChecklistStep";
-export * from "./steps/MediaGalleryStep";
-export * from "./steps/PricingDurationStep";
-export * from "./steps/ReviewSubmitStep";
