+++ src/pages/Watchlist.tsx
@@ -5,7 +5,7 @@ import { api } from "convex/_generated/api";
 import { Heart, Loader2 } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
-import { AuctionCard } from "@/components/auction";
+import { AuctionCard } from "@/components/auction/AuctionCard";
 import { LoadingIndicator } from "@/components/LoadingIndicator";
 import {
   PAGINATION_INITIAL_ITEMS,
