--- src/pages/admin/dialogs/index.ts
@@ -1,8 +1,0 @@
-/**
- * Admin dialog components for user and auction management operations
- * Extracted from monolithic AdminDialogs.tsx for better maintainability
- */
-
-export { KycReviewDialog } from "./KycReviewDialog";
-export { BulkActionDialog } from "./BulkActionDialog";
-export { PromoteAdminDialog } from "./PromoteAdminDialog";
